pyproject.toml
[project]
name = "graph-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.25.1",
    "azure-keyvault-secrets>=4.10.0",
    "azure-messaging-webpubsubservice>=1.3.0",
    "azure-mgmt-resource>=24.0.0",
    "dotenv>=0.9.9",
    "mem0ai>=1.0.0",
    "sendgrid>=6.12.5",
]
